feat!: rename package tree
BREAKING CHANGE: package tree is renamed

Fixes #313
Addresses #311 excluding repository instructions
