+++ pom.xml
@@ -16,7 +16,7 @@ limitations under the License.
 -->
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <groupId>com.alexandriasoftware.swing</groupId>
+    <groupId>com.github.rhwood.jsplitbutton</groupId>
     <artifactId>jsplitbutton</artifactId>
     <version>2.0.0</version>
     <packaging>jar</packaging>
@@ -56,9 +56,8 @@ limitations under the License.
     <developers>
         <developer>
             <name>Randall Wood</name>
-            <email>randall.h.wood@alexandriasoftware.com</email>
-            <organization>Randall Wood DBA Alexandria Software</organization>
-            <organizationUrl>https://alexandriasoftware.com</organizationUrl>
+            <organization>Randall Wood</organization>
+            <organizationUrl>https://rhwood.github.io</organizationUrl>
         </developer>
     </developers>
 
@@ -254,7 +253,7 @@ limitations under the License.
                 <configuration>
                     <archive>
                         <manifestEntries>
-                            <Automatic-Module-Name>com.alexandriasoftware.swing.jsplitbutton</Automatic-Module-Name>
+                            <Automatic-Module-Name>com.github.rhwood.jsplitbutton</Automatic-Module-Name>
                         </manifestEntries>
                     </archive>
                 </configuration>
