+++ src/main/resources/com/github/rhwood/jsplitbutton/splitbutton_16.png
