+++ src/main/resources/com/github/rhwood/jsplitbutton/splitbutton_32.png
